# aspnetcore-server-swagger/src/IO.Swagger/Controllers/CoronaStatusApi.cs
/*
 * Floripa vrs Corona - Phonebook
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 0.0.4
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */

using System;
using System.Collections.Generic;
using Microsoft.AspNetCore.Mvc;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using System.ComponentModel.DataAnnotations;
using IO.Swagger.Attributes;
  
using Microsoft.AspNetCore.Authorization;
using IO.Swagger.Models;

namespace IO.Swagger.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class CoronaStatusApiController : ControllerBase
    { 
        /// <summary>
        /// 
        /// </summary>
        
        /// <param name="id1"></param>
        /// <response code="200">successful operation</response>
        [HttpGet]
        [Route("//phonebook/contacts/id/{id1}/journal")]
        [ValidateModelState]
        [SwaggerOperation("PhonebookContactsIdId1JournalGet")]
        [SwaggerResponse(statusCode: 200, type: typeof(List<CoronaStatus>), description: "successful operation")]
        public virtual IActionResult PhonebookContactsIdId1JournalGet([FromRoute][Required]string id1)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(List<CoronaStatus>));

            string exampleJson = null;
            exampleJson = "[ {\n  \"id\" : \"id\",\n  \"timestamp\" : 6,\n  \"status\" : \"status\"\n}, {\n  \"id\" : \"id\",\n  \"timestamp\" : 6,\n  \"status\" : \"status\"\n} ]";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<List<CoronaStatus>>(exampleJson)
            : default(List<CoronaStatus>);
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// 
        /// </summary>
        
        /// <param name="id1"></param>
        /// <param name="id2"></param>
        /// <response code="200">successful operation</response>
        [HttpGet]
        [Route("//phonebook/contacts/id/{id1}/journal/id/{id2}")]
        [ValidateModelState]
        [SwaggerOperation("PhonebookContactsIdId1JournalIdId2Get")]
        [SwaggerResponse(statusCode: 200, type: typeof(CoronaStatus), description: "successful operation")]
        public virtual IActionResult PhonebookContactsIdId1JournalIdId2Get([FromRoute][Required]string id1, [FromRoute][Required]string id2)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(CoronaStatus));

            string exampleJson = null;
            exampleJson = "{\n  \"id\" : \"id\",\n  \"timestamp\" : 6,\n  \"status\" : \"status\"\n}";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<CoronaStatus>(exampleJson)
            : default(CoronaStatus);
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// 
        /// </summary>
        
        /// <param name="id1"></param>
        /// <param name="body"></param>
        /// <response code="200">successful operation</response>
        [HttpPatch]
        [Route("//phonebook/contacts/id/{id1}/journal")]
        [ValidateModelState]
        [SwaggerOperation("PhonebookContactsIdId1JournalPatch")]
        [SwaggerResponse(statusCode: 200, type: typeof(CoronaStatus), description: "successful operation")]
        public virtual IActionResult PhonebookContactsIdId1JournalPatch([FromRoute][Required]string id1, [FromBody]CoronaStatus body)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(CoronaStatus));

            string exampleJson = null;
            exampleJson = "{\n  \"id\" : \"id\",\n  \"timestamp\" : 6,\n  \"status\" : \"status\"\n}";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<CoronaStatus>(exampleJson)
            : default(CoronaStatus);
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// 
        /// </summary>
        
        /// <param name="id1"></param>
        /// <param name="body"></param>
        /// <response code="200">successful operation</response>
        [HttpPost]
        [Route("//phonebook/contacts/id/{id1}/journal")]
        [ValidateModelState]
        [SwaggerOperation("PhonebookContactsIdId1JournalPost")]
        [SwaggerResponse(statusCode: 200, type: typeof(CoronaStatus), description: "successful operation")]
        public virtual IActionResult PhonebookContactsIdId1JournalPost([FromRoute][Required]string id1, [FromBody]CoronaStatus body)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(CoronaStatus));

            string exampleJson = null;
            exampleJson = "{\n  \"id\" : \"id\",\n  \"timestamp\" : 6,\n  \"status\" : \"status\"\n}";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<CoronaStatus>(exampleJson)
            : default(CoronaStatus);
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// 
        /// </summary>
        
        /// <param name="id1"></param>
        /// <param name="body"></param>
        /// <response code="200">successful operation</response>
        [HttpPut]
        [Route("//phonebook/contacts/id/{id1}/journal")]
        [ValidateModelState]
        [SwaggerOperation("PhonebookContactsIdId1JournalPut")]
        [SwaggerResponse(statusCode: 200, type: typeof(CoronaStatus), description: "successful operation")]
        public virtual IActionResult PhonebookContactsIdId1JournalPut([FromRoute][Required]string id1, [FromBody]CoronaStatus body)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(CoronaStatus));

            string exampleJson = null;
            exampleJson = "{\n  \"id\" : \"id\",\n  \"timestamp\" : 6,\n  \"status\" : \"status\"\n}";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<CoronaStatus>(exampleJson)
            : default(CoronaStatus);
            //TODO: Change the data returned
            return new ObjectResult(example);
        }
    }
}