# aspnetcore-server-swagger/src/IO.Swagger/wwwroot/swagger-original.json
{
  "swagger" : "2.0",
  "info" : {
    "version" : "0.0.4",
    "title" : "Floripa vrs Corona - Phonebook"
  },
  "host" : "67.205.157.240:8080",
  "basePath" : "/",
  "paths" : {
    "/phonebook/contacts" : {
      "get" : {
        "tags" : [ "Contact" ],
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Contact"
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Contact" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Contact"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Contact"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "Contact" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Contact"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Contact"
            }
          }
        }
      },
      "patch" : {
        "tags" : [ "Contact" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Contact"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Contact"
            }
          }
        }
      }
    },
    "/phonebook/contacts/id/{id1}" : {
      "get" : {
        "tags" : [ "Contact" ],
        "parameters" : [ {
          "name" : "id1",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Contact"
            }
          }
        }
      }
    },
    "/phonebook/contacts/id/{id1}/journal" : {
      "get" : {
        "tags" : [ "CoronaStatus" ],
        "parameters" : [ {
          "name" : "id1",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/CoronaStatus"
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "CoronaStatus" ],
        "parameters" : [ {
          "name" : "id1",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/CoronaStatus"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/CoronaStatus"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "CoronaStatus" ],
        "parameters" : [ {
          "name" : "id1",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/CoronaStatus"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/CoronaStatus"
            }
          }
        }
      },
      "patch" : {
        "tags" : [ "CoronaStatus" ],
        "parameters" : [ {
          "name" : "id1",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/CoronaStatus"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/CoronaStatus"
            }
          }
        }
      }
    },
    "/phonebook/contacts/id/{id1}/journal/id/{id2}" : {
      "get" : {
        "tags" : [ "CoronaStatus" ],
        "parameters" : [ {
          "name" : "id1",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "id2",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/CoronaStatus"
            }
          }
        }
      }
    }
  },
  "definitions" : {
    "Contact" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "string"
        },
        "phone" : {
          "type" : "string"
        },
        "name" : {
          "type" : "string"
        },
        "age" : {
          "type" : "integer"
        },
        "journal" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/CoronaStatus"
          }
        }
      },
      "example" : {
        "journal" : [ {
          "id" : "id",
          "timestamp" : 6,
          "status" : "status"
        }, {
          "id" : "id",
          "timestamp" : 6,
          "status" : "status"
        } ],
        "phone" : "phone",
        "name" : "name",
        "id" : "id",
        "age" : 0
      }
    },
    "CoronaStatus" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "string"
        },
        "timestamp" : {
          "type" : "integer",
          "format" : "int64"
        },
        "status" : {
          "type" : "string"
        }
      },
      "example" : {
        "id" : "id",
        "timestamp" : 6,
        "status" : "status"
      }
    }
  }
}